app/requirements.txt
fastapi
fastapi-cli
ollama
pydantic
groq
pytest
httpx
requests
pytest-rerunfailures
weave
groq
python-dotenv
PyYAML